Added ability to choose to get RMS error instead of just total % error from predictions.
This is to see whether the reason the accuracy isn't growing in some test cases is because of A[L] changing but not significantly enough for the predictions to change.
a.k.a. 0.2 for a prediction of 1 is incorrect so in the next pass it could have changed to 0.21 but the prediction (0 or 1) is still incorrect so no change in err% will be noticed in the old way.
For now using RMS(L2) instead of the actual loss function just to help with my intuition. Later, might change to add L(y, y_hat) as well.